+++ NEWS.md
@@ -3,7 +3,7 @@
 ## Change in 1.3.4
 
 - Add funtions to generate `ggplot` figures: `plotChromatograms()` and
-  `plotChromatogramsOverlay()`. 
+  `plotChromatogramsOverlay()`.
 
 ## Changes in 1.3.3