+++ NEWS.md
@@ -2,8 +2,8 @@
 
 ## Change in 1.3.4
 
-- Add funtions to generate `ggplot` figures: `plotChromatograms()` and
-  `plotChromatogramsOverlay()`.
+- Add funtions to generate `ggplot` figures: `ggplotChromatograms()` and
+  `ggplotChromatogramsOverlay()`.
 
 ## Changes in 1.3.3